Moving connection logic to the connetion handler
