+++ connections/aleo/connection_test.go
@@ -1,1 +1,13 @@
+/* Copyright (c) 2021 Forte Labs
+ * All Rights Reserved.
+ *
+ * NOTICE:  All information contained herein is, and remains the property of
+ * Forte Labs and their suppliers, if any.  The intellectual and technical
+ * concepts contained herein are proprietary to Forte Labs and their suppliers
+ * and may be covered by U.S. and Foreign Patents, patents in process, and are
+ * protected by trade secret or copyright law. Dissemination of this information
+ * or reproduction of this material is strictly forbidden unless prior written
+ * permission is obtained from Forte Labs.
+ */
+
 package aleo
